Reenable free variables
Treat free variables in propositions without modal operators as
constants denoting actual objects.  Don't decide propositions with both
free variables and modal operators in them.

+++ perl/upl.pm
@@ -2,28 +2,53 @@ use strict;
 use IPC::Open2;
 
 sub decide;
+sub complexity;
+sub query;
 
 my $out;
 my $in;
 my $pid = 0;
 
-sub decide
+
+sub query
 {
   $pid = open2( $out, $in, 'upl --query' ) unless $pid;
-  print $in ( @_ ? $_[0] : $_ );
-  print $in "\n?\n";
+  (my $statement, my $query_type) = @_;
+  $statement = $_ if not defined $statement;
+  $query_type = '' if not defined $query_type;
+  print $in $statement;
+  print $in "\n$query_type?\n";
   my $result = <$out>;
-  die unless defined $result;
+  die "failed to get result from upl.exe" unless defined $result;
   $result =~ s/[\r\n]//g;
   return $result;
 }
 
+sub decide
+{
+  return &query;
+}
+
+sub complexity
+{
+  return int query( @_ ? $_[0] : $_, 'c' );
+}
+
+sub is_sentence
+{
+  my $result = query( @_ ? $_[0] : $_, 's' );
+  return $result =~ /Valid/;
+}
+
 END
 {
-  close $in;
-  close $out;
-  kill -9, $pid;
-  waitpid( $pid, 0 );
+  if ( $pid )
+  {
+    close $in;
+    close $out;
+    kill -9, $pid;
+    waitpid( $pid, 0 );
+  }
 }
 
 return 1;
